[FIX] web_editor: fix handles preventing editing snippets
Steps to reproduce the bug:

- In Website edit mode.
- Drag and drop a "Badge" snippet into the footer.
- Bug: it is not possible to edit the text of the badge or change its
icon.

The problem occurs because the handles of the snippet overlay cover the
entire surface of the snippet, and the pointer event is intercepted by
these handles.

Note that the fix does not need to be made before version 17 because,
prior to that, the handles have a smaller height and therefore do not
cause this bug.

Their height was changed with this commit [1].

[1]: https://github.com/odoo/odoo/commit/70f723a9f78c406746f5409f6a89cbe93ad21580

task-3927962

closes odoo/odoo#165657

Signed-off-by: Colin Louis (loco) <loco@odoo.com>